Fix Windows tun read handling
It's too fragile to depend on the 'new_pkt' check, and only do the ReadFile()
call when we allocate a new packet buffer. Keep track of the pending state
directly in the Windows-specific code, instead.

This should fix the fact that it gets out of sync when we get an error.
Such as the ERROR_MORE_DATA errors we get because vpnc-script-win.js
doesn't set the MTU properly yet...

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>