android: Introduce new "mirror fetcher" shell script
fetch.sh will try downloading our dependencies from several different
mirrors, in case the project's main website is down.  Add known-good SHA1
hashes to android/Makefile so fetch.sh can detect bad downloads.

This is intended to work in a similar manner to the OpenWRT downloader
(which unfortunately was too closely coupled with their build system to
reuse as-is).

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>