Move DTLS secret initialisation to openconnect_setup_dtls()
This was added in commit ec2408e5 ("dtls: Align new-tunnel rekey behavior
with Cisco clients"), and was causing the NetworkManager authentication
dialog to crash because it was calling openconnect_random() before the SSL
library was initialised by openconnect_init_ssl().

The auth dialog didn't need it anyway. Move it to openconnect_setup_dtls()
where it belongs.

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>