Fix DTLS master secret generation (harder)
Kevin points out that doing it in openconnect_setup_dtls(), as in commit
fa2d8b4a, is too late. It needs to be present by the time we do the CSTP
connection, since it's passed in the headers — even if we subsequently
end up not doing DTLS at all.

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>