Attempt to make sense of GSSAPI mess
This should build fairly cleanly everywhere. FreeBSD will whine at us if
we use <gssapi.h> but Solaris requires it. Most platforms have krb5-config
but on Solaris it doesn't do 'krb5-config gssapi', and some older platforms
lack it.

If neither --without-gssapi nor --with-gssapi is specified to configure,
try to find it but just fall back to building without.

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>