Add strndup() compat function for OSX
The comment in connect_https_socket() suggests that I didn't use strndup()
there because Solaris lacks it. I blithely went ahead and used it in the
NTLM code though, and Solaris 11 *does* seem to have it. (My Solaris 10 VM
no longer boots, and I can't bring myself to care.)

But OSX (10.6.8 Snow Leopard) also seems to lack strndup(). So add a
compat implementation of it rather than just trying to remember not to
use it.

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>