Handle non-ASCII tun device names under Windows
Most of the calls here can remain resolutely ASCII-only, since the registry
key names aren't going to contain arbitrary characters. The cosmetic device
name, on the other hand, does need to be obtained as "Unicode" and converted
to UTF-8.

We can't just use WideCharToMultiByte here because it doesn't exist on
older versions of Windows (yay!).

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>