http: Fix CSD https requests
If request_body is NULL, buf_error(request_body) will return -ENOMEM to
indicate an allocation failure.  This is a good thing if we just called
buf_alloc(), but it is a bad thing if we aren't actually passing in a
request_body pointer.

The CSD GET requests pass in a NULL request_body and request_body_type:

    result = do_https_request(vpninfo, "GET", NULL, NULL, &form_buf, 0);

so this change is needed to restore proper CSD operation.

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>