Fixes for buf_append_utf16le()
Make it set buf_error() when it returns -EINVAL, to ensure callers will
catch the problem. Previously, ntlm_set_string_binary() and  sspi_setup()
would not have noticed.

Also, NUL-terminate the buffer. Not accounting for it, but just putting
two zero bytes after the string so that it can be used as a
NUL-terminated (wchar_t *). Not entirely sure how sspi_setup() was
working before this, except perhaps pure luck.

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>