auth: Refactor stoken form handling
Split the devid/pass form from the PIN form, to allow the use of new "v3"
tokens from the command line.  On newer libstoken versions, query the
PIN status and token interval time; use this information to figure out
whether to request a PIN for concatenation (ala hard tokens) and whether
to use a 30- or 60-second delta in response to the next tokencode prompt.

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>