Fix endless loop when automatic NTLM auth fails
We never *do* actually fall back from automatic to manual NTLM auth since
our criteria for using each are mutually exclusive. So arguably we could
simplify things somewhat and just set the state to AUTH_FAILED if the SSO
version fails. But let's preserve the option to fall back later, if we do
start prompting for a password on demand.

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>