Import AM_ICONV implementation
We *have* to have config.rpath in our own directory, for reasons which
escape me. Usually it would be installed by gettextize or gnulib-tool
but we're not *using* either of those, so just pull in the AM_ICONV bits
manually.

We don't want to ship just config.rpath and *not* the M4 files because
that way we might end up with mismatching versions.

I hate autohell.

(Files taken from Fedora gettext 0.19.2-5.fc21)

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>