Fix thread-unsafe setenv() usage
Since moving the tunnel handling into libopenconnect too, we've been
unsafely using setenv() before spawning vpnc-script. We mustn't do that
in a multi-threaded environment.

So store up all the variables in a list, and apply them after fork(). Or
on Windows, build up the environment block to pass to CreateProcess()
manually.

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>