Allow console output from vpnc-script on Windows
In commit f3b06b62a ("use CreateProcess instead of system to run scripts.")
we used the CREATE_NO_WINDOW flag to CreateProcess() which avoids creating
a new console window for the vpnc-script, in the GUI client.

That had the side-effect of also not allowing vpnc-script to output to
the *existing* console when invoked from the command-line client. Check
whether there's a console or not and only use CREATE_NO_WINDOW if there
isn't, and both should have the behaviour they desire.

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>