Stop calling setenv() from JNI code
If there's ever an environment that's likely to be multi-threaded, it's
Java. So we really can't use setenv() directly. Instead, add a new
openconnect_set_csd_environ() function and set the environment variables
only after fork().

With a hack to make it look for a $TMPDIR setting because that gets used
*before* forking, for the download location for the script.

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>