Do not call vpn_progress() after forking
It's suboptimal that we only write to stderr, but calling the progress_fn
from a fork is going to cause all kinds of problems. Either we leave it
as it is or we come up with a saner way of passing the message back.

Perhaps a *new* 'child_write_progress' function which the caller can
provide, which does IPC or something. For now, stderr will suffice.

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>