Try harder to find PKCS#11 key to match a given cert
CKA_LABEL might not match, but CKA_ID should. So if the label (of the cert)
was in the original search criteria, but the ID wasn't, then as a last
resort try *dropping* CKA_LABEL from the URI and adding the CKA_ID of the
cert.

Now it works to use a PIV card with something simple like
openconnect -c 'pkcs11:object=Certificate%20for%20PIV%20Authentication'

Also allow ;object-type=cert to be specified for the certificate. We're
clever enough to work out that a key needs to have that changed...

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>