Do compression context setup *after* negotiation rather than before
It doesn't make much sense to preemptively set up contexts for every type
of compression we support, before negotiating with the server. Allocate
it afterwards instead.

Yes, it means we can't silently drop compression support if the allocation
fails, but really was everything going to turn out fine in that case anyway?

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>