Always output specific IP address in authentication results
We previously kept the hostname if there was only *one* result returned
from DNS — on the assumption that if there was more than one, we might get
a different IP each time and we needed to remember *which* one we connected
to, but if there's only one it wouldn't change.

With trick DNS setups like vpn.intel.com, that isn't true. It can change
every time. So just set vpninfo->unique_hostname *every* time.

As noted in the comment, there is a problem if we're connecting through
a proxy, because we don't *know* which IP address we connected to. We
probably ought to do the DNS lookup locally and then ask the proxy to
connect to the specific IP address... except that we can't always rely
on that working.

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>