Start separating protocol-specific methods from generic VPN support
It looks like we can understand the Juniper oNCP protocol and it fits fairly
well with the OpenConnect model — you first authenticate and are rewarded
with a cookie, and then you make the actual connection to the HTTPS server
and *can* transport data over that, or ideally you have a UDP transport
(in this case ESP) instead.

So start factoring out the methods which are specific to the VPN protocol,
with a view to making Juniper support possible too.

Starting with cstp_bye().

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>