Add shell of Juniper support
The API to enable this is a dirty hack for now; we'll work it out
properly later. It's entirely feasible that the auth UI will want to
be given the raw HTML to render it, since we really *can* be given
arbitrary web pages (with Java and JavaScript) and we're expected to
fill the forms in.

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>