Render U+002D HYPHEN-MINUS in manual page where needed
In groff, plain '-' is a hyphen, interpreted according to context but not
necessarily the plain ASCII U+002D HYPHEN-MINUS. So it's common to use \-
instead... but that's defined as a minus sign, and in HTML it's emitted
as U+2212 MINUS SIGN — still not what we want.

cf. https://bugzilla.redhat.com/show_bug.cgi?id=1173619
and https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=159872#54

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>