Add LZO decompression support
It doesn't look like we get to negotiate this. If the server is configured
to do compression, it *will* compress packets it sends to us via ESP.

So unless we want to eschew ESP entirely when compression is offered, we
need at least a decompressor which is LGPL-compatible (unlike liblzo2).

Thankfully. libavutil happens to have one that we can steal. It doesn't
look that hard to do a compressor for it too.

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>