Calculate TOTP/HOTP codes for ourselves
No released version of liboath supports SHA256/SHA512. At this point it
looks like more trouble than it's worth, since generating an HMAC is
fairly trivial and we already do it in various places already.

This is simpler than adding conditional support for the new function
that liboath will eventually add to allow us to specify the HMAC
algorithm we want it to use.

(SHA1 only for now; we need to work out how the user asks for it too)

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>