cstp: Add X-AnyConnnect-* mobile headers on CONNECT request
The AnyConnect mobile clients send these headers, and as it so happens,
ocserv uses them to detect whether to apply "mobile-idle-timeout" or
"idle-timeout" to the connection.  Currently ocserv applies "idle-timeout"
to all mobile and non-mobile openconnect clients because we aren't sending
the necessary headers.  This commit fixes that.

This only has an effect when "--os android" or "--os apple-ios" is used,
or if openconnect_set_mobile_info() is called.  The headers will not be
sent for traditional Linux/OSX/... clients.

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>