Use canonical representation of 'sí' in Spanish translation
We ask the user to *type* this in response to a question about whether
to accept an SSL certificate which could not be validated. It helps if
it actually *matches* what the user is likely to type.

For some reason the translation used U+0301 COMBINING ACUTE ACCENT after
a simple ASCII 'i' character. Change it to use U+00ED LATIN SMALL LETTER
I WITH ACUTE.

Strictly speaking, we should probably be doing a normalised compare, as
we could get from libicu's ucol_strcoll() function (or CompareString()
on Windows). But this should suffice for now.

http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=792073

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>