Do not shutdown tun if it isn't running
This significantly improves the error reporting when the tun setup fails.

We were getting a lot of spew from a failed vpnc-script 'disconnect'
invocation, which was obscuring the real error messages.

Especially in the "you need to be running as root" case, which a lot of
unsophisticated users see and which we've carefully tailored to be readable,
we *really* don't want it to be obscured.

With this fix, we do still have a final 'Unknown error; exiting' line of
stderr output which we didn't previously have before the MTU detection.
I'll see if I can eliminate that too.

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>