Remove oncp_https_submit() function
We want to revamp this code flow a little to allow a callback to the GUI
with the HTML page; this split out oncp_https_submit() function is not
helping. So just inline it within its only caller, and we'll sort it out
from there. This commit should have no behavioural changes at all.

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>