Move Juniper check_cookie_success() before HTML parsing
If we're going to call out to the application to render HTML, let's
restructure things so we can avoid our own HTML parsing completely.

Now the last thing we need this for is TNCC. Which *maybe* we could
expect a real HTML renderer to handle anyway?

We need to work out how an HTML renderer should submit its own HTTP
requests — does it use libopenconnect, or its own transport? What
about cookies?

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>