Allow OC_CMD_PAUSE to abort connection attempts
Currently OC_CMD_PAUSE does not abort connection attempts in all cases.
One example of this can be seen in openssl.c:openconnect_open_https(),
which can get stuck forever if the socket hangs.

Change the helper function so that either OC_CMD_PAUSE or OC_CMD_CANCEL
exits the mainloop; the main difference is whether it is handled as a
temporary or permanent disconnection.

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>
Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>