Add --passtos option to copy TOS/TCLASS from VPN packets
This allows prioritised queuing of outbound packets. It is only of local
significance (and importance) as it will influence queueing on the CPE
which is typically the only place where this will be in effect. And the
most effective place as the CPE is usually the bottleneck where all
applications compete for limited upstream bandwidth.

SPs do set the DSCP to 0 anyway at the trust boundary (which is the next
hop from the CPE). Same goes for large corporations which also either
reset the DSCP or have it set according to their policy, not the user's.

It is implemented as an 'opt-in' using the --passtos command line switch
in accordance with the OpenVPN implementation

Signed-off-by: Ralph Schmieder <ralph.schmieder@gmail.com>
Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>