Fix retry on tun fd when it isn't a Linux tun device
If we get ENOBUFS, EAGAIN or EWOULDBLOCK, keep the packet on the queue
and retry.

The kernel allegedly returns ENOMEM when its queue for the tun device is
full, although I can't reproduce that now even with a while(1) loop writing
the same packet to the device over and over again. So maybe it's changed.

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>