Add translations for 'Enter PKCS#8 pass phrase:'
Even without knowing the languages and the rules about vowel and plurals
and the like, I can guess that this is going to be *remarkably* similar
to the translations for 'Enter PKCS#12 pass phrase:'.

I know there's a slight possibility that languages exist in which the
number really does make a different to the rest of the sentence, but
I can live with that.

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>