Don't start the GPST tunnel if ESP is connecting or connected (FIXME timing)
The GP ESP tunnel doesn't work if we've started the GP SSL tunnel. For now,
I just close the HTTPS connection and avoid opening it if ESP is connecting
or connected.

This isn't a good solution, however. It only gives a small and unpredictable
time window for ESP to detect a connection before GPST will take over.