Temporarily replace Juniper esp_send_probes() with GP ESP version
To test and start the ESP tunnel, the GP Windows client sends 3× ping/ICMP
request from the client address to the server address.  The server won't
respond unless the first 16 bytes of the following magic payload are
included; the remainder of the payload seems to be arbitrary.

    char magic[56] = "monitor\x00\x00pan ha " /* required bytes */
                     "0123456789:;<=>? !\"#$%&\'()*+,-./\x10\x11\x12\x13\x14\x15\x16\x18";

Putting together IPv4 ICMP packets "by hand" is not that fun. References:

  * http://stackoverflow.com/questions/8290046/icmp-sockets-linux
  * https://austinmarton.wordpress.com/2011/09/14/sending-raw-ethernet-packets-from-a-specific-interface-in-c-on-linux/