gpst: Clean up some 'const char *' that shouldn't be const
... and the associated forced casts to non-const on free()

Although actually... given that we aren't doing the string expansion
thing, perhaps the 'var' returned by xmlnode_get_text() *could* be
const in this case, and just return node->content. Most callers do
seem to have to free it instead of really keeping it, and the few
that want to keep it can use strdup(). Dan?

Signed-off-by: David Woodhouse <dwmw2@infradead.org>