add --request-ip option to explicitly request a specific IPv4 addresses
This behavior is supported by AnyConnect (CONNECT with X-CSTP-Address header)
and by GlobalProtect (POST /ssl-vpn/getconfig.esp with preferred-ip form field).

Currently, this option is only a request. OpenConnect will print an error
message, but will not abort, if the server assigns a different IPv4 address.

I did not implement the corresponding behavior for IPv6 because I do not
have a way to test it right now.